% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infor_table.R
\name{infor_table}
\alias{infor_table}
\title{Information table.}
\usage{
infor_table(x, snp, covar_n, MAF_value, model)
}
\arguments{
\item{x}{An output from continuous/binary/survival model.}

\item{snp}{Single SNP name.}

\item{covar_n}{Covariate names.}

\item{MAF_value}{A minimum value of minor allele frequency.}

\item{model}{Model type.}
}
\value{
Information table. If linear or binary model is chosen, it returns estimated coefficients, 
confidence interval and P value. If survival model is chosen, it returns hazard ratio, confidence interval and P value.
}
\description{
\code{infor_table} returns information table of estimated coefficients/hazard ratio, confidence interval and P value.
}
