% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAF.R
\name{MAF}
\alias{MAF}
\title{Select SNP by MAF.}
\usage{
MAF(snp, gender, male, MAF_v, data)
}
\arguments{
\item{snp}{SNP name.}

\item{gender}{gender variable.}

\item{male}{male information.}

\item{MAF_v}{minimum MAF value.}

\item{data}{a dataset.}
}
\description{
\code{MAF} returns SNPs with higher MAF than default value.
}
